+++ test/azure/legacy/Expected/AcceptanceTests/Head/head/aio/_patch.py
@@ -11,7 +11,6 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from typing import List
-
 __all__ = []  # type: List[str]  # Add all objects you want publicly available to users at this package level